fix translation of wx priority values to the standard Unix ones in SetPriority()
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@48537 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/unix/threadpsx.cpp
@@ -1311,7 +1311,7 @@ void wxThread::SetPriority(unsigned int prio)
 
             // map wx priorites WXTHREAD_MIN_PRIORITY..WXTHREAD_MAX_PRIORITY
             // to Unix priorities 20..-20
-            if ( setpriority(PRIO_PROCESS, 0, -(2*prio)/5 + 20) == -1 )
+            if ( setpriority(PRIO_PROCESS, 0, -(2*(int)prio)/5 + 20) == -1 )
             {
                 wxLogError(_("Failed to set thread priority %d."), prio);
             }
